fix(merge): PR #11 review followups — uint32, afs guards, count cache
Addresses the PR #11 review notes ahead of the v0.4 tag:

- **uint32 _pgen_row** instead of int64. Halves the per-pvar memory
  footprint (336 MB instead of 672 MB on an 84M-variant panel — ~6%
  of the existing pvar DataFrame), matches pgenlib's native variant_idx
  type, and pd.merge's float64 round-trip remains lossless for any
  realistic variant_ct (well under 2^53).

- **Wire `afs` into both startup checks** (`check_max_alleles` and
  `check_pvar_pgen_row_count_consistent`). Previously afs ran neither
  — multi-allelic input would SIGSEGV inside pgenlib's read_range
  exactly as it does in merge / validate / inspect, and a mis-paired
  triplet would silently corrupt frequencies.

- **Eliminate the count_raw_variants double-scan.**
  `check_pvar_pgen_row_count_consistent` now returns the raw .pvar
  data-line count it computed, and accepts an optional pre-computed
  `n_pvar` to skip the scan when a caller already has it. merge_cmd
  captures the return value and reuses it for the descriptor
  n_variants population in step 7; inspect_cmd passes its n_pre_filter
  through to the guardrail. Saves ~30-60s of redundant I/O at
  84M-variant panel scale.

- **Fix misleading test docstring** in TestNonCanonicalReadByteIdentity
  (was referencing `--on-collision first`; the test actually uses
  distinct IIDs and doesn't need a collision policy).

+++ src/pgen_samplebind/afs.py
@@ -164,7 +164,9 @@ def compute_afs(
     # Use the `_pgen_row` column read_pvar stamps before filtering so the
     # read_list call lands on the right .pgen rows even when biallelic
     # indels were dropped from the pandas DataFrame.
-    original_indices = pvar.loc[keep_mask, "_pgen_row"].to_numpy().astype(np.int64)
+    # uint32 from read_pvar — pgenlib's native variant_idx type; the
+    # `int(...)` calls at the read_range/read boundaries upcast cleanly.
+    original_indices = pvar.loc[keep_mask, "_pgen_row"].to_numpy()
 
     # 4. Streaming aggregation. Per-(variant, population) accumulators:
     #    alt_count[v, p]    = Σ over samples in p: g     (diploid)
